+++ app/src/main/java/com/manosprojects/themoviedb/domain/usecase/MarkMovieAsFavouriteUC.kt
@@ -4,13 +4,13 @@ import com.manosprojects.themoviedb.domain.repository.MoviesRepository
 import javax.inject.Inject
 
 interface MarkMovieAsFavouriteUC {
-    fun execute(movieId: Int, isFavourite: Boolean)
+    fun execute(movieId: Long, isFavourite: Boolean)
 }
 
 class MarkMovieAsFavouriteUCImpl @Inject constructor(
     private val moviesRepository: MoviesRepository
 ) : MarkMovieAsFavouriteUC {
-    override fun execute(movieId: Int, isFavourite: Boolean) {
+    override fun execute(movieId: Long, isFavourite: Boolean) {
         moviesRepository.markMovieAsFavourite(movieId = movieId, isFavourite = isFavourite)
     }
