+++ app/src/main/java/com/manosprojects/themoviedb/features/home/contract/HomeContract.kt
@@ -10,7 +10,7 @@ data object HomeContract {
     data class State(val showLoading: Boolean, val movies: List<HomeMovieModel>) : UiState
 
     sealed interface Effect : UiEffect {
-        data class NavigateToMovie(val movieId: Int) : Effect
+        data class NavigateToMovie(val movieId: Long) : Effect
         data class ShowSnack(val message: String) : Effect
     }
