+++ app/src/main/java/com/manosprojects/themoviedb/navigation/NavigationActions.kt
@@ -3,6 +3,6 @@ package com.manosprojects.themoviedb.navigation
 import androidx.navigation.NavController
 
 
-fun NavController.navigateToMovieDetails(movieId: Int) {
+fun NavController.navigateToMovieDetails(movieId: Long) {
     navigate(Route.MovieDetails(movieId = movieId.toString()).destination)
 }
